Actually disable BnB when there are preset inputs
Summary:
We don't want to use BnB when there are preset inputs because there
is some weirdness with making that work with using the KnapsackSolver
as the fallback. Currently we say that we haven't used bnb when
there are preset inputs, but we don't actually disable BnB. This fixes
that.

Test that BnB is not used when there are preset inputs

This is a backport of Core PR12694

Depends on D3364

Test Plan:
  make check

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: markblundeberg

Differential Revision: https://reviews.bitcoinabc.org/D3378