Don't use the functional test arguments in the tmp directory name
Summary:
The `tmpdir` of a functional test is constructed using the whole test
command line, including arguments.
This breaks the assumption that the name ends with `.py` in the case
there are arguments (such as `--usecli` or `--mineblock`), introduces
spaces and other undesirable characters in the directory name, and makes
it harder to determine the tmpdir name from the test suite.
This diffs changes this behavior to only use the test name to build the
tmpdir name. Note that there is no collision because the unique
`portseed` is also part of the name (this is left unchanged).

Test Plan:
  ./test/functional/test_runner.py wallet_txn_clone --nocleanup
Look at the directory structure.
Before the diff:
  <test_runner tmpdir>/'wallet_txn_clone.py --mineblock_<N>'/
  <test_runner tmpdir>/wallet_txn_clone_<N+1>/
After the diff:
  <test_runner tmpdir>/wallet_txn_clone_<N>/
  <test_runner tmpdir>/wallet_txn_clone_<N+1>/

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3151