Remove unknown version warning from UpdateTip
Summary:
Since Antminer S9s now are able to do overt asicboost, there are many unknown
versions on mainnet headers.  This is triggering warnings in the debug.log and it
is not apparent that this is meaningful any longer.  Additionally, the BIP9 code
has been removed for activating any softforks based on the version bits anyways.

Depends on D3381

Test Plan:
  make check

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix, schancel, jasonbcox, Fabien

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox, Fabien

Subscribers: markblundeberg

Differential Revision: https://reviews.bitcoinabc.org/D2788